+++ module-apps/application-calendar/widgets/RepeatAndReminderItem.cpp
@@ -15,29 +15,44 @@ namespace gui
         activeItem = false;
         setEdges(RectangleEdge::None);
         setMinimumSize(style::window::default_body_width, style::window::calendar::item::repeatAndReminder::height);
+        setMargins(gui::Margins(style::margins::small, 0, 0, 0));
+
+        std::vector<gui::Image *> squareImages;
+        for (int i = 0; i < 2; i++) {
+            auto square = new gui::Image();
+            square->set("Rectangle");
+            square->setAlignment(gui::Alignment{gui::Alignment::Horizontal::Left, gui::Alignment::Vertical::Center});
+            squareImages.push_back(square);
+        }
 
         hBox = new HBox(this, 0, 0, 0, 0);
         hBox->setEdges(RectangleEdge::None);
 
         repeatVBox = new VBox(hBox, 0, 0, 0, 0);
-        repeatVBox->setMinimumSize(style::window::default_body_width / 2,
+        repeatVBox->setMinimumSize(style::window::default_body_width / 2 - style::margins::small / 2,
                                    style::window::calendar::item::repeatAndReminder::height);
         repeatVBox->setEdges(RectangleEdge::None);
 
         reminderVBox = new VBox(hBox, 0, 0, 0, 0);
-        reminderVBox->setMinimumSize(style::window::default_body_width / 2,
+        reminderVBox->setMinimumSize(style::window::default_body_width / 2 - style::margins::small / 2,
                                      style::window::calendar::item::repeatAndReminder::height);
         reminderVBox->setEdges(RectangleEdge::None);
 
-        repeatTitle = new gui::Label(repeatVBox, 0, 0, 0, 0);
+        repeatHBox = new HBox(repeatVBox, 0, 0, 0, 0);
+        repeatHBox->setMinimumSize(style::window::default_body_width / 2 - style::margins::small / 2,
+                                   style::window::calendar::item::repeatAndReminder::title_label_h);
+        repeatHBox->setMargins(
+            gui::Margins(0, 0, 0, style::window::calendar::item::repeatAndReminder::title_label_margin));
+        repeatHBox->setEdges(RectangleEdge::None);
+        repeatHBox->addWidget(squareImages[0]);
+
+        repeatTitle = new gui::Label(repeatHBox, 0, 0, 0, 0);
         repeatTitle->setMinimumSize(style::window::calendar::item::repeatAndReminder::description_w,
                                     style::window::calendar::item::repeatAndReminder::title_label_h);
-        repeatTitle->setMargins(
-            gui::Margins(0, 0, 0, style::window::calendar::item::repeatAndReminder::title_label_margin));
+        repeatTitle->setMargins(gui::Margins(style::margins::small, 0, 0, 0));
         repeatTitle->setEdges(RectangleEdge::None);
         repeatTitle->setFont(style::window::font::small);
         repeatTitle->setAlignment(gui::Alignment{gui::Alignment::Horizontal::Left, gui::Alignment::Vertical::Center});
-        repeatTitle->setLineMode(true);
         repeatTitle->activeItem = false;
 
         repeat = new gui::Label(repeatVBox, 0, 0, 0, 0);
@@ -48,15 +63,21 @@ namespace gui
         repeat->setAlignment(gui::Alignment{gui::Alignment::Horizontal::Left, gui::Alignment::Vertical::Center});
         repeat->activeItem = false;
 
-        reminderTitle = new gui::Label(reminderVBox, 0, 0, 0, 0);
-        reminderTitle->setMinimumSize(style::window::default_body_width / 2,
-                                      style::window::calendar::item::repeatAndReminder::title_label_h);
-        reminderTitle->setMargins(
+        reminderHBox = new HBox(reminderVBox, 0, 0, 0, 0);
+        reminderHBox->setMinimumSize(style::window::default_body_width / 2 - style::margins::small / 2,
+                                     style::window::calendar::item::repeatAndReminder::title_label_h);
+        reminderHBox->setMargins(
             gui::Margins(0, 0, 0, style::window::calendar::item::repeatAndReminder::title_label_margin));
+        reminderHBox->setEdges(RectangleEdge::None);
+        reminderHBox->addWidget(squareImages[1]);
+
+        reminderTitle = new gui::Label(reminderHBox, 0, 0, 0, 0);
+        reminderTitle->setMinimumSize(style::window::calendar::item::repeatAndReminder::description_w,
+                                      style::window::calendar::item::repeatAndReminder::title_label_h);
+        reminderTitle->setMargins(gui::Margins(style::margins::small, 0, 0, 0));
         reminderTitle->setEdges(RectangleEdge::None);
         reminderTitle->setFont(style::window::font::small);
         reminderTitle->setAlignment(gui::Alignment{gui::Alignment::Horizontal::Left, gui::Alignment::Vertical::Center});
-        reminderTitle->setLineMode(true);
         reminderTitle->activeItem = false;
 
         reminder = new gui::Label(reminderVBox, 0, 0, 0, 0);